Refactored to clippy spec
